requirements.txt
# Requirements for Flappy Bird Dark Edition web deployment
pygame>=2.6.0
pygbag==0.7.1
pillow>=10.0.0
opencv-python-headless>=4.8.0